// ui/src/app/kanbas/model/UserJwt.ts
export class UserJwt{
  "iss": string;
  "email": string;
  "email_verified": boolean;
  "name": string;
  "picture": string;
  "given_name": string;
  "family_name": string;
}

// "iss": "https://accounts.google.com",
//   "nbf": 1669300979,
//   "aud": "991781410138-tvl1u9dqgm7a271ov0eg54mcrpps79ve.apps.googleusercontent.com",
//   "sub": "112810122988321427859",
//   "email": "fredy.hernandez-orozco@capgemini.com",
//   "email_verified": true,
//   "azp": "991781410138-tvl1u9dqgm7a271ov0eg54mcrpps79ve.apps.googleusercontent.com",
//   "name": "Fredy Hernandez Orozco",
//   "picture": "https://lh3.googleusercontent.com/a/ALm5wu3TGZp0TfbHq6Z8jGBkF41lKEFQe1sOcSKE0-XJ=s96-c",
//   "given_name": "Fredy",
//   "family_name": "Hernandez Orozco",
//   "iat": 1669301279,
//   "exp": 1669304879,
//   "jti": "cfa3acae5cac8863735a6f6b5172c85433350381"